各種バグの修正
