+++ app/assets/stylesheets/application.scss
@@ -10,12 +10,14 @@
  * files in this directory. Styles in this file should be added after the last require_* statement.
  * It is generally better to create a new file per style scope.
  *
- *= require flatpickr
  *= require_tree .
  *= require_self
  */
 
 @import "bootstrap/scss/bootstrap";
+@import "flatpickr/dist/flatpickr";
+@import "flatpickr/dist/themes/material_blue.css";
+
 
 .graph-container {
     @extend .container-fluid;
