+++ app/javascript/packs/application.js
@@ -1,0 +1,13 @@
+require("@rails/ujs").start()
+require("turbolinks").start()
+require("@rails/activestorage").start()
+require("channels")
+
+require("bootstrap/dist/js/bootstrap")
+
+require('chart.js')
+
+require('flatpickr')
+require('flatpickr/dist/l10n/ja')
+
+require("./graghs")
