+++ app/views/layouts/application.html.erb
@@ -4,7 +4,7 @@
     <title>Safekeepingmoney</title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-    <meta name="viewport" content="width=devise-width,initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
   </head>
@@ -16,5 +16,6 @@
    <% end %>
   <div class="graph-container">
     <%= yield %>
+  </div>
   </body>
 </html>
